requirements.txt
requests>=2.31.0
pydantic>=2.6.0
supabase>=2.3.0
python-dotenv>=1.0.0
neo4j>=5.15.0
psycopg2-binary>=2.9
